removed redundant functions
